ANA-1196: Add optional `valueFormatter` prop
Add an optional `valueFormatter` that can be passed to format
aggregation results before they are rendered.
This is only done if the currently selected aggregator outputs a
value in the same unit as the source data.

Update `Dropdown` component so that it closes itself after an
option is selected.
Update uses of `Dropdown` so that when a new aggregator is selected,
we clear the selections for the aggregation targets.

Add source maps to the webpack dev server config
for easier debugging when running locally.

Update version number and regenerate package-lock with new
version number.