CON-94: Add declaration file
Add a TypeScript declaration file so that we can consume this package
in our typescript projects without having to provide our own typings.
This will hopefully mean we are more likely to keep the types in sync with
the JS and will mean we don't need to do workarounds to get the types
for this package bundled into our other libraries.

Update the build script to build to a `dist` directory rather than the
root so that we can use the `fields` option in package.json to
include everything inside of `dist` instead of having to list all the
relevant files individually.
Update this script to also copy the new declaration file into the build
output.
Add the `types` field to `package.json` to make sure consumers
can pick the types up.